[mod] moves list components to list folder
