--- src/components/categories/PredictionsVisibility/index.ts
@@ -1,1 +1,0 @@
-export { PredictionVisibility } from "./PredictionVisibility";
