+++ src/components/list-items/CategoryPredictionListItem/index.ts
@@ -1,0 +1,1 @@
+export { CategoryPredictionListItem } from "./CategoryPredictionListItem";
