+++ src/components/list-items/index.ts
@@ -13,3 +13,5 @@ export { SaveImageListItem } from "./SaveImageListItem";
 export { SaveProjectListItem } from "./SaveProjectListItem";
 export { SegmenterExecListItem } from "./SegmenterExecListItem";
 export { SegmenterArchitectureSettingsListItem } from "./ArchitectureSettingsListItem";
+export { CategoryItem } from "./CategoryItem";
+export { CategoryPredictionListItem } from "./CategoryPredictionListItem";
