+++ src/components/lists/SegmenterList/SegmenterList.tsx
@@ -9,7 +9,7 @@ import { useDialog, useDialogHotkey } from "hooks";
 import { ImportTensorflowModelDialog } from "components/dialogs";
 import { CollapsibleList } from "components/common/styled-components/CollapsibleList";
 import { SaveFittedModelDialog } from "components/dialogs";
-import { CategoriesList } from "components/categories/CategoriesList";
+import { CategoriesList } from "components/lists/CategoriesList";
 import { SegmenterExecListItem } from "components/list-items";
 
 import { selectCreatedAnnotationCategories } from "store/data";
