+++ src/components/lists/SegmenterList/index.ts
@@ -1,0 +1,1 @@
+export { SegmenterList } from "./SegmenterList";
