+++ src/components/lists/index.ts
@@ -1,0 +1,6 @@
+export { FileList } from "./FileList";
+export { ClassifierList } from "./ClassifierList";
+export { CategoriesList } from "./CategoriesList";
+export { ImageList } from "./ImageList";
+export { ClearAnnotationsGroup } from "./ClearAnnotationsGroup";
+export { SegmenterList } from "./SegmenterList";
