+++ src/components/menus/index.ts
@@ -4,3 +4,4 @@ export { OpenProjectMenu } from "./OpenProjectMenu";
 export { OpenExampleProjectMenuItem } from "./OpenExampleProjectMenuItem";
 export { ImageMenu } from "./ImageMenu";
 export { SaveImageMenu } from "./SaveImageMenu";
+export { CategoryItemMenu } from "./CategoryItemMenu";
