--- src/components/segmenter/SegmenterList/index.ts
@@ -1,1 +1,0 @@
-export { SegmenterList as ClassifierList } from "./SegmenterList";
